+++ example/test/widget_test.dart
@@ -1,9 +1,11 @@
 import 'package:example/main.dart' as ma;
+import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:idkit/idkit.dart';
 
 void main() {
   test('Function Test', () {
-    // ma.timeFormatHandle();
-    ma.thousandsHandle();
+    ma.desensitizedMethod();
+    ma.phoneAndBankFormatMethod();
   });
 }