+++ lib/idkit.dart
@@ -4,4 +4,6 @@ export 'src/function/date/func_date.dart';
 export 'src/function/thousands/func_thousands.dart';
 export 'src/widget/widget_gap.dart';
 export 'src/widget/choice/widget_choice.dart';
-export 'src/function/color/func_color_tool.dart';
+export 'src/function/color/func_color.dart';
+export 'src/function/desensitized/fun_desensitized.dart';
+export 'src/function/format/func_format_num.dart';